backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
google-genai==1.47.0
python-dotenv==1.2.1
pillow==11.3.0
python-multipart==0.0.20
httpx==0.28.1
websockets>=12.0
